+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 ## Release Notes
 
+`1.0.12`
+
+- Fixed wrong order for `Methods.getAllDeclaredInHierarchy(Class, Predicate)`.
+- Changed `Comparables` constructor to throw `UnsupportedOperationException` to prevent instantiation.
+- Upgraded `pitest-maven` to `1.22.0`.
+- Upgraded `jacoco-maven-plugin` to `0.8.14`.
+- Upgraded `maven-gpg-plugin` to `3.2.8`.
+- Upgraded `central-publishing-maven-plugin` to `0.9.0`.
+- Moved `Reflection.findSubclass` method to `Classes`.
+- Added `Serializable` to `ObjectConverter` to allow serialization of converters.
+- Added `Nullables.Chain.thenNotNull` to chain null checks.
+
+---
+
 `1.0.11`
 
 - Added `Methods.getCallerMethodName` without a supplier to get the caller method name.